Limit pgalloc.tryPopulate() to 1 GB.
Syzkaller came up with workloads that fallocate(2) 1 TB. The host mlock(2)
or madvise(2) syscalls end up hanging for multiple minutes in such situations
causing the watchdog to mark the calling goroutine as stuck.

The intent behind tryPopulate() is to prepopulate shmem pages on the host to
avoid potential future page faults. This is an optimization. It is not part of
fallocate(2) semantics.

Most systems have on the order of GBs of RAM. It doesn't make sense to populate
unlimited pages as they will keep being swapped after a point.

Reported-by: syzbot+4aa3d6d42b063a11c850@syzkaller.appspotmail.com
PiperOrigin-RevId: 544362539